feat(auth): add authentication page
Add an authentication page, react-query will be used for caching data
from the API.  openapi-typescript will be used for making the API
type-safe on the client-side.

The authentication page contacts the API to obtain a token, which is
stored using context.  When this token is set, the user is considered
"logged in".